requirements-dev.txt
# Зависимости для разработки и тестирования

# Telegram Bot API
python-telegram-bot>=21.0.0

# Whisper
openai-whisper>=20231117

# NeMo
nemo-toolkit[collections]>=1.0.0

# Ollama
ollama>=0.1.0

# LLM
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99

# Тестирование
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
requests-mock>=1.11.0

# Линтинг и форматирование
flake8>=6.1.0
mypy>=1.7.0
bandit>=1.7.6
black>=23.11.0
isort>=5.12.0
yamllint>=1.32.0
pre-commit>=3.4.0

# Утилиты